fix(quality): implement reviewed ReSharper remediation
Implement the 2026-09-10 ReSharper remediation plan with 1,363 fixes,
956 reviewed exceptions and seven owned runtime-integration follow-ups.
Correct deep rollback restoration and exception parameter names; retain
public DTO contracts, runtime boundary guards and protected replay data.

Add pinned all-framework InspectCode execution and a fail-closed SARIF
regression gate with an auditable original-to-current disposition ledger.

Validated with ReSharper CLI 2026.2.1 (zero new or unresolved in-scope
findings), 3,226 Debug/Release tests, 24 smoke tests, six locked replay
goldens, 19 tooling tests and a verified netstandard2.1 plugin refresh.

+++ src/ProjectAegis.Delegation.Tests/Projection/ContactProvenanceProjectionTests.cs
@@ -53,7 +53,7 @@ public sealed class ContactProvenanceProjectionTests
 
         var snapshot = ContactProvenanceProjection.Project(
             log,
-            currentSimTick: 1UL + (ulong)ContactProvenanceProjection.DefaultStaleThresholdTicks + 1);
+            currentSimTick: 1UL + ContactProvenanceProjection.DefaultStaleThresholdTicks + 1);
 
         Assert.That(snapshot.Contacts[0].Freshness, Is.EqualTo(ContactProvenanceFreshness.Stale));
         Assert.That(snapshot.Contacts[0].AgeTicks, Is.EqualTo((ulong)ContactProvenanceProjection.DefaultStaleThresholdTicks + 1));
